Move OCP processing-related methods to correct source file
